refactor: clean featuredSlider component and move Card component to ProductSlider folder
