+++ src/components/ProductSlider/ProductSlider.module.scss
