+++ src/components/ProductSlider/index.js
@@ -1,0 +1,2 @@
+import ProductSlider from "./ProductSlider";
+export default ProductSlider;
